Pistol for the Player
Added a player weapon, the ability to shoot it, player projectiles, and a small fix to make looking around consistent regardless of framerate. The weapon has an ammunition count which determines if it can shoot or not. It can be "equipped" or "unequipped" by enabling or disabling the weapon's PlayerPistol script. The bullets try to determine if they hit an enemy by checking if an object they hit has the "Enemy" tag.

A prefab for the player object can be found in the prefabs folder. Everything listed here has been implemented into the "Player Walk Test" scene to test out.