P2 Alpha Changes (#36)
* Added and improved enemies

* Added two new weapons, random spread, aim-down-sights, and more

Two new weapons - melee axe and thrown grenades - which show off the different types of weapons I can implement with what I've scripted. Switch between weapons with the mouse scrollwheel or the 1 and 2 keys.
Weapons can have random bullet spread. Horizontal and Vertical spread can be changed separately.
Aim-down-sights can be added for specific weapons, affecting random spread.
Also made some simple models for a couple things to differentiate them.

* Update More Weapons Test.unity

* Enemies can also be hit by trigger colliders

* updated player prefab

* QOL changes,

* csharp stuff

* uistuff scene

* add utils

* Update ProjectSettings.asset

* Update ProjectSettings.asset

* setup health bar

* fix nullable enable

* fix nullable enable

* stamina bar

* weapon ui

* cleanup stuff

* accidently invoked the ammochanged event in update instead of on shoot

* hitmarkers

* adjust size of bars

* kill hitmarker

* objective ui

* oops

* hide stamina bar if in vehicle

* move hud elements to hud object, start work on a pause menu

* blur behind ui elements

* options menu

doesn't pause the game

* crosshair

* variable spread

* update ammo counter when picking up ammo

* adjust vignette

* hide crosshair lines if there is no spread

* bullet fx

* improve bullet trail

* adjust look of bullet trail

* null check

* adjust smoke trail, move bullet disabling to its own method

* New pistol model, cooking grenades, and car pickups

Grenade weapon now uses the "PlayerThrowGrenade" script, as a child to "PlayerPistol". Hold the attack button to cook grenades and release to throw. The pistol looks a little better now. Refuel and repair pickups are now a thing and work exactly as you might expect.

* Armor and Health pickups

Currently, armor works as a percent damage negation. Usually that is not how armor works in games, so I'll probably change it in a future commit.

* Armor as bonus health

PlayerHealth now has an armorPoints value, which should absorb received damage so long as it is greater than 0. Armor pickups add to this value. Armor as a percent of damage negation still exists as well, in case we wish to use it.

* Forgot to give the player health script to the grenade

whoops

* basic grenade vfx (needs smoke/etc)

* improve explosion fx

* adjust explosion particles

* interpolate grenade rigid body

* Enemy animations (#33)

* Added animations for Zombie enemies

* Fixed the animations a little and added sounds for the zombies

---------

Co-authored-by: rks32 <rks32njit.edu>
Co-authored-by: Jonathan Malave <95556326+jonathanmal1119@users.noreply.github.com>

* World building (#34)

* Car Obstacles Reintegrated.

* New Gas Station

* Map Changes and billboard added.

* BIllboard Promo Added

* More world Building

* More Small Details and Vehcile changes

* Objective Created

* Small Enemy Bugs

* UI and FX (#35)

* ticking ui for cooking grenades, also adjust grenade physics

* make bullet trail slightly thinner

* use particle ribbon for smoke trail

* adjust smoke trail material

* adjust trails again

* improve explosion fx

* adjust shockwave

* cleanup shadergraph

* turn canvas into prefab

* Update Player Pistol Bullet Projectile.prefab

* hide dot

* QOL and integration

* Final last minute changes.

---------

Co-authored-by: rks32 <rks32njit.edu>
Co-authored-by: nlm25 <nlm25@njit.edu>
Co-authored-by: Dyvinia <dulana57@gmail.com>
Co-authored-by: rks32 <144377760+rks32@users.noreply.github.com>